+++ typings/index.d.ts
@@ -445,8 +445,8 @@ declare module 'klasa' {
 		public readonly client: KlasaClient;
 		public readonly gateway: Gateway;
 		public readonly id: string;
+		public readonly synchronizing: boolean;
 		private _existsInDB: boolean;
-		private _syncStatus?: Promise<void>;
 
 		public get(key: string): any;
 		public get<T>(key: string): T;
@@ -456,8 +456,11 @@ declare module 'klasa' {
 
 		public reset(key?: string | string[], options?: ConfigurationResetOptions): Promise<ConfigurationUpdateResult>;
 		public reset(key?: string | string[], guild?: KlasaGuild, options?: ConfigurationResetOptions): Promise<ConfigurationUpdateResult>;
-		public update(key: ObjectLiteral<any>, guild?: GuildResolvable): Promise<ConfigurationUpdateResult>;
+		public update(key: ObjectLiteral<any>, options?: ConfigurationUpdateOptions): Promise<ConfigurationUpdateResult>;
+		public update(key: ObjectLiteral<any>, guild?: GuildResolvable, options?: ConfigurationUpdateOptions): Promise<ConfigurationUpdateResult>;
+		public update(key: string, value: any, options?: ConfigurationUpdateOptions): Promise<ConfigurationUpdateResult>;
 		public update(key: string, value: any, guild?: GuildResolvable, options?: ConfigurationUpdateOptions): Promise<ConfigurationUpdateResult>;
+		public update(key: string[], value: any[], options?: ConfigurationUpdateOptions): Promise<ConfigurationUpdateResult>;
 		public update(key: string[], value: any[], guild?: GuildResolvable, options?: ConfigurationUpdateOptions): Promise<ConfigurationUpdateResult>;
 		public list(message: KlasaMessage, path: SchemaFolder | string): string;
 		public resolveString(message: KlasaMessage, path: SchemaPiece | string): string;
@@ -487,13 +490,13 @@ declare module 'klasa' {
 		public defaultSchema: object;
 		public readonly resolver: SettingResolver;
 		public readonly cache: Collection<string, Configuration>;
+		public readonly syncQueue: Collection<string, Promise<Configuration>>;
 
 		public get(input: string | number, create?: boolean): Configuration;
-		public sync(input?: object | string): Promise<Configuration>;
-		public sync(download: boolean): Promise<null>;
+		public sync(input?: string[]): Promise<Gateway>;
+		public sync(input: string | { id?: string, name?: string }): Promise<Configuration>;
 		public getPath(key?: string, options?: GatewayGetPathOptions): GatewayGetPathResult | null;
 
-		private _download(): Promise<void>;
 		private _resolveGuild(guild: GuildResolvable): KlasaGuild;
 		private _shardSync(path: string[], data: any, action: 'add' | 'delete' | 'update'): Promise<void>;
 
@@ -540,7 +543,7 @@ declare module 'klasa' {
 
 		public register(name: string, schema?: object, options?: GatewayDriverRegisterOptions): this;
 		public init(): Promise<void>;
-		public sync(): Promise<Array<null>>;
+		public sync(input?: string[]): Promise<Array<Gateway>>;
 
 		public toJSON(): GatewayDriverJSON;
 		public toString(): string;
@@ -1274,7 +1277,7 @@ declare module 'klasa' {
 		ownerID?: string;
 		permissionLevels?: PermissionLevels;
 		pieceDefaults?: KlasaPieceDefaults;
-		prefix?: string;
+		prefix?: string | string[];
 		preserveConfigs?: boolean;
 		providers?: KlasaProvidersOptions;
 		readyMessage?: (client: KlasaClient) => string;